requirements.txt
langchain
pinecone-client
python-dotenv
streamlit
sentence-transformers 
pinecone-client 
huggingface-hub 
python-dotenv
langchain-community